requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy
pillow
tqdm

# CLIP
git+https://github.com/openai/CLIP.git

# Perceptual metric
lpips

# Image metrics
scikit-image

# Video processing
opencv-python

# Utilities
scipy